[admin] Split admin script to make it testable
